Rebase fixes from develop
